refactor: 공통으로 사용하던 주문 상태 클래스 패키지 이동
