+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -18,7 +18,6 @@ import javax.persistence.Table;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 
 @Table(name = "orders")
